pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-trial-env"
version = "1.0.0"
description = "AI Agent Environment for Clinical Research Coordination"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "requests>=2.28.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
datagen = [
    "google-generativeai>=0.4.0",
]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["clinical_trial_env*", "server*"]
